refactor(editor): move the CodeMirror wiring into a hook
Fixing the reconfigure took WorksheetEditor from 8 hooks to 16 and past
Fallow's cognitive-complexity threshold, which counts a React hook as a
decision point. The gate is right: the component was doing the editor's
whole configuration and its rendering.

`use-worksheet-editor` now owns all of it — the compartment, the effects,
the callbacks and the extension list — and returns props that are already
identity-stable, so the rule that keeps the editor from reconfiguring
lives in one module with the comment explaining why. The component is
back to JSX.

One ref holds the latest props rather than one ref per callback, which is
both fewer hooks and less to keep in step; the keymap and the handlers
read `latest.current` at the moment they fire.
